refactor: 拆分 test_caption_numbering → utils/test_text + rules/test_caption
- tests/utils/test_text.py (186行): FromRoman, FromChineseNum, ParseCaptionText
  零外部依赖，只 import wordformat.utils 的三个函数
- tests/rules/test_caption.py (778行): 合并题注格式测试 + 编号集成测试
- 删除 test_caption_numbering.py
- 943 passed, 0 failed
